requirements.txt
annotated-types==0.7.0
anyio==4.7.0
chardet==5.2.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fastapi==0.115.6
fonttools==4.55.3
h11==0.14.0
idna==3.10
kiwisolver==1.4.7
loguru==0.7.3
lxml==5.3.0
matplotlib==3.9.3
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
reportlab==4.2.5
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
